requirements.txt
requests>=2.31.0
selenium>=4.15.0
google-cloud-storage>=2.14.0
pandas>=2.1.0
python-dotenv>=1.0.0 